+++ src/pages/RegistrationPage/RegistrationPage.tsx
